requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.22.0,<=1.24.3
plotly>=5.15.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
ccxt>=4.0.0
ta>=0.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
scikit-learn>=1.2.0
tensorflow>=2.12.0
joblib>=1.2.0
faiss-cpu>=1.7.4
scipy>=1.9.0
typing_extensions>=4.8.0